+++ lib/tailwind_ui.dart
@@ -6,5 +6,6 @@ export 'src/application_ui/elements/avatars/circular_avatar.dart';
 export 'src/application_ui/elements/avatars/circular_avatar_with_notification.dart';
 export 'src/application_ui/elements/avatars/circular_avatar_with_placeholder_icon.dart';
 export 'src/application_ui/elements/avatars/circular_avatar_with_placeholder_initials.dart';
+export 'src/application_ui/elements/avatars/profile_tile.dart';
 export 'src/application_ui/elements/avatars/rounded_avatar.dart';
 export 'src/application_ui/elements/avatars/rounded_avatar_with_notification.dart';